simplified event reporter files split + event logger reporter
