+++ lib/metrics/reporter/events/event-reporter-base.ts
@@ -1,4 +1,8 @@
-import { EventReporterOptions } from "./event-reporter-options";
+import { Logger } from "../logger";
+
+export interface EventReporterOptions {
+  log?: Logger;
+}
 
 export abstract class EventReporterBase<O extends EventReporterOptions> {
   protected readonly options: O;
