--- lib/metrics/reporter/events/event-reporter-options.ts
@@ -1,5 +1,0 @@
-import { Logger } from "../logger";
-
-export interface EventReporterOptions {
-  log?: Logger;
-}
